Updated IndexKeyGenerator location.
