+++ geo/s2index.go
@@ -27,11 +27,8 @@ import (
 	"github.com/dgraph-io/dgraph/x"
 )
 
-// KeyGenerator generates keys used by geo spatial indexes.
-type KeyGenerator struct{}
-
 // IndexKeys creates the keys to use in the index for the given data assuming that it is geo data.
-func (k KeyGenerator) IndexKeys(attr string, data []byte) ([]string, error) {
+func IndexKeys(data []byte) ([]string, error) {
 	// Try to parse the data as geo type.
 	v, err := types.GeoType.Unmarshaler.FromBinary(data)
 	if err != nil {
@@ -41,12 +38,12 @@ func (k KeyGenerator) IndexKeys(attr string, data []byte) ([]string, error) {
 	if !ok {
 		log.Fatalf("Unexpected type from the unmarshaler.")
 	}
-	return IndexKeys(g)
+	return IndexKeysFromGeo(g)
 }
 
-// IndexKeys returns the keys to be used in a geospatial index for the given geometry. If the
+// IndexKeysFromGeo returns the keys to be used in a geospatial index for the given geometry. If the
 // geometry is not supported it returns an error.
-func IndexKeys(g types.Geo) ([]string, error) {
+func IndexKeysFromGeo(g types.Geo) ([]string, error) {
 	cu, err := indexCells(g)
 	if err != nil {
 		return nil, err
